🔨 refactor: history 파일 분리
