--- src/widgets/history/styles/book/content_container/Award.css
@@ -1,113 +1,0 @@
-.award__container {
-  display: flex;
-  flex-direction: column;
-
-  width: 100%;
-  height: 100%;
-
-  justify-content: center;
-  align-items: center;
-
-  gap: 1.04vmax;
-}
-
-.award__content {
-  display: flex;
-  flex-direction: column;
-
-  width: 100%;
-  height: 100%;
-
-  justify-content: space-between;
-  gap: 0;
-}
-
-.award__year-group {
-  display: flex;
-  flex-direction: column;
-
-  gap: 0.26vmax;
-}
-
-.award__year {
-  font-weight: 700;
-  font-size: var(--text-md);
-  color: var(--black);
-}
-
-.award__ul {
-  display: flex;
-  flex-direction: column;
-  list-style: none;
-}
-
-.award__ul li {
-  display: flex;
-
-  gap: 1.93%;
-
-  font-size: var(--text-sm);
-  font-weight: 400;
-
-  color: var(--black);
-}
-
-.award__ul li::before {
-  content: '•';
-  flex-shrink: 0;
-
-  font-size: var(--text-sm);
-}
-
-@media (max-width: 1024px) {
-  .award__container {
-    gap: 1.3vw;
-  }
-
-  .award__content {
-    align-items: flex-start;
-    gap: 0;
-  }
-
-  .award__year-group {
-    gap: 0.39vw;
-  }
-
-  .award__year {
-    font-size: 1.17vw;
-  }
-
-  .award__ul li {
-    gap: 0.39vw;
-
-    font-size: 1.04vw;
-  }
-
-  .award__ul li::before {
-    font-size: 1.04vw;
-  }
-}
-
-@media (max-width: 767px) {
-  .award__container {
-    gap: clamp(9px, 2.67vw, 13px);
-  }
-
-  .award__year-group {
-    gap: clamp(2px, 0.8vw, 4px);
-  }
-
-  .award__year {
-    font-size: clamp(8px, 2.4vw, 12px);
-  }
-
-  .award__ul li {
-    gap: clamp(3px, 1vw, 5px);
-
-    font-size: clamp(8px, 2.13vw, 11px);
-  }
-
-  .award__ul li::before {
-    font-size: clamp(8px, 2.13vw, 11px);
-  }
-}
